solve mapper problem

+++ sportscenter-Backend/src/main/java/com/ecommerce/sportscenter/entity/OrderAggregate/Order.java
@@ -31,7 +31,7 @@ public class Order {
     @Column(name="Sub_Total")
     private Double subTotal;
     @Column(name="Delivery_Fee")
-    private Long deliveryFee;
+    private Double deliveryFee;
     @Enumerated(EnumType.STRING)
     @Column(name="Order_Status")
     private OrderStatus orderStatus = OrderStatus.Pending;
