+++ sportscenter-Backend/src/main/java/com/ecommerce/sportscenter/model/OrderDto.java
@@ -1,13 +1,12 @@
 package com.ecommerce.sportscenter.model;
 
-import java.time.LocalDateTime;
-
 import com.ecommerce.sportscenter.entity.OrderAggregate.ShippingAddress;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class OrderDto {
     private Integer basketId;
     private ShippingAddress shippingAddress;
-    private Long subTotal;
-    private Long deliveryFee;
+    private Double subTotal;
+    private Double deliveryFee;
     private LocalDateTime orderDate;
 }
