+++ sportscenter-Backend/src/main/java/com/ecommerce/sportscenter/model/OrderResponse.java
@@ -1,14 +1,13 @@
 package com.ecommerce.sportscenter.model;
 
-import java.time.LocalDateTime;
-
 import com.ecommerce.sportscenter.entity.OrderAggregate.OrderStatus;
 import com.ecommerce.sportscenter.entity.OrderAggregate.ShippingAddress;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 
 @Data
 @Builder
@@ -18,8 +17,8 @@ public class OrderResponse {
     private Integer id;
     private Integer basketId;
     private ShippingAddress shippingAddress;
-    private Long subTotal;
-    private Long deliveryFee;
+    private Double subTotal;
+    private Double deliveryFee;
     private Double total;
     private LocalDateTime orderDate;
     private OrderStatus orderStatus;
